new homepage prototype
Just finished prototyping a new homepage. I think the /follow page is
really awkward looking and useless so I wanted to get rid of it and put
the latest/upcoming location right under the title in that white
outlined box then also get rid of the "map" link if the map is already
visible and then put social media on the homepage and also get rid of
the way that the navigation overlaps the map, because it was kinda cool
but also kinda illegible and when you're trying to interact with map
pins and they’re behind the header it wouldn’t work. Then, the map was
really fucked up on phones, like it worked but when you clicked on a
point the scroll would break, so I hid the map from the front page on
iPhones, and instead will link to the map as a separate page. I think
it makes a lot more sense now and is starting to look nicer too. Also
made the map darker to match the Rodi gray, roughly, and made the map
popovers nicer looking, with full bleed images, and fixing the
type/links.